refactor(db): Reorganize CountryRepository batch operations
This commit refactors the batch operations for the `CountryRepository` to improve code organization and reduce file size. The batch operation implementations and their corresponding tests have been moved into a dedicated directory structure.

Components affected:
- `banking-db-postgres`:
  - Moved `country_repository_batch_impl.rs` to `repository/person/country_repository/`
  - Updated `lib.rs` and `mod.rs` to reflect the new file structure.
  - Relocated `country_batch_operations_test.rs` to be co-located with the implementation.

Benefits:
- **Improved Code Organization**: Batch operations are now organized in a dedicated directory, improving clarity.
- **Enhanced Maintainability**: Smaller, more focused files are easier to understand and maintain.
- **Better Test Co-location**: Tests are now located with the code they are testing, making them easier to find and run.
