+++ banking-db-postgres/src/repository/person/mod.rs
@@ -9,4 +9,4 @@ pub mod location_repository_batch_impl;
 pub mod location_repository_impl;
 pub mod person_repository_batch_impl;
 pub mod person_repository_impl;
-pub mod country_repository_batch_impl;
+pub mod country_repository;
