+++ banking-db-postgres/tests/suites/person/location_batch_operations_test.rs
@@ -8,7 +8,7 @@ use uuid::Uuid;
 use rust_decimal::Decimal;
 use std::str::FromStr;
 
-use crate::suites::person::country_batch_operations_test::setup_test_country;
+use crate::suites::person::helpers::setup_test_country;
 use crate::suites::person::country_subdivision_batch_operations_test::setup_test_country_subdivision;
 use crate::suites::person::locality_batch_operations_test::setup_test_locality;
 use crate::suites::test_helper::setup_test_context;
